Changed shouts to be displayed on status bar level
