+++ Source/Message.swift
@@ -23,7 +23,7 @@ public struct Announcement {
   public var duration: NSTimeInterval
   public var action: (() -> Void)?
 
-  public init(title: String, subtitle: String? = nil, image: UIImage? = nil, duration: NSTimeInterval = 2, action: (() -> Void)? = nil) {
+  public init(title: String, subtitle: String? = nil, image: UIImage? = nil, duration: NSTimeInterval = 20, action: (() -> Void)? = nil) {
     self.title = title
     self.subtitle = subtitle
     self.image = image
